dom/promise/tests/test_promise_retval.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1436276.
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1436276.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="file_promise_retval_tests.js"></script>
  <script type="application/javascript">
  /** Test for Bug 1436276. **/
  SimpleTest.waitForExplicitFinish();

  function verifyPromiseGlobal(p, global, msg) {
    // SpecialPowers.Cu.getGlobalForObject returns a SpecialPowers wrapper for
    // the actual global.  We want to grab the underlying object.
    var globalWrapper = SpecialPowers.Cu.getGlobalForObject(p);
    is(SpecialPowers.unwrap(globalWrapper), global,
       msg + " should come from " + global.label);
  }

  function expectedExceptionGlobal(global) {
    // We should end up with an exception from "global".
    return global;
  }

  function getPromise(global, arg) {
    return global.TestFunctions.passThroughPromise(arg);
  }

  addLoadEvent(function() {
    frames[0].label = "child";
    SpecialPowers.pushPrefEnv({set: [['dom.expose_test_interfaces', true]]},
                              runPromiseRetvalTests.bind(undefined,
							 SimpleTest.finish));
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1323324">Mozilla Bug 1323324</a>
<p id="display"></p>
<div id="content" style="display: none">
  <!-- A subframe so we have another global to work with -->
  <iframe></iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
